+++ website/projects/scimaps/src/app/app.component.ts
@@ -4,7 +4,8 @@ import { ActivatedRoute, Params, Router, RouterOutlet } from '@angular/router';
 import { Select, Store } from '@ngxs/store';
 import { Observable, of, Subscription } from 'rxjs';
 import { debounceTime, distinctUntilChanged, map, switchMap } from 'rxjs/operators';
-import { drawerInOut, slideWithTransform } from './constants/drawer.animations';
+import { drawerInOut } from './constants/drawer.animations';
+import { slideWithTransform } from './constants/route.animations';
 import { isAssetURL, isExternalURL } from './constants/utils';
 import { SetAppState } from './core/actions/app.actions';
 import { PageState } from './core/state/page/page.state';
